refactor(runtime): expose lifecycle status
