+++ console/src/regions/LeftRail/LeftRail.tsx
@@ -1,6 +1,12 @@
 import { useEffect, useState } from "react";
 import { actions, useStore } from "../../store";
-import type { ProcessInfo } from "../../types";
+import type {
+  ProcessInfo,
+  ProcessStatus,
+  RuntimeChildInfo,
+  RuntimeReadiness,
+  RuntimeStatusSnapshot,
+} from "../../types";
 import { AudioStrip } from "./AudioStrip";
 import { ProcessCard } from "./ProcessCard";
 import "./leftrail.css";
@@ -25,7 +31,14 @@ export function LeftRail() {
           className={"mic-toggle-btn" + (micActive ? " active" : "")}
           onClick={() => actions.setMicActive(!micActive)}
         >
-          {micActive ? (<><span className="mic-live-dot" />open mic live</>) : "turn mic on"}
+          {micActive ? (
+            <>
+              <span className="mic-live-dot" />
+              open mic live
+            </>
+          ) : (
+            "turn mic on"
+          )}
         </button>
       </div>
 
@@ -44,21 +57,106 @@ export function LeftRail() {
 
 function useProcesses(): ProcessInfo[] {
   const [list, setList] = useState<ProcessInfo[]>([]);
+
   useEffect(() => {
     let alive = true;
+
     async function poll() {
       try {
-        const res = await fetch("/api/process/status");
-        if (alive && res.ok) setList(await res.json());
+        const [processes, runtime] = await Promise.all([
+          fetchProcessStatus(),
+          fetchRuntimeStatus(),
+        ]);
+        if (alive) setList(attachRuntimeChildren(processes, runtime));
       } catch {
-        // supervisor not ready — leave the last good list
+        // Supervisor not ready: leave the last good list visible.
       }
     }
+
     poll();
     const id = setInterval(poll, POLL_MS);
-    return () => { alive = false; clearInterval(id); };
+    return () => {
+      alive = false;
+      clearInterval(id);
+    };
   }, []);
+
   return list;
+}
+
+async function fetchProcessStatus(): Promise<ProcessInfo[]> {
+  const res = await fetch("/api/process/status");
+  if (!res.ok) throw new Error("process status unavailable");
+  return res.json();
+}
+
+async function fetchRuntimeStatus(): Promise<RuntimeStatusSnapshot | null> {
+  try {
+    const res = await fetch("/runtime/status");
+    if (!res.ok) return null;
+    return res.json();
+  } catch {
+    return null;
+  }
+}
+
+function attachRuntimeChildren(
+  processes: ProcessInfo[],
+  runtime: RuntimeStatusSnapshot | null,
+): ProcessInfo[] {
+  const app = processes.find((process) => process.name === "kaguya_app");
+  if (!app || app.status === "stopped" || !runtime?.manage_processes) {
+    return processes;
+  }
+
+  const children = runtimeChildren(runtime);
+  return processes.map((process) =>
+    process.name === "kaguya_app" ? { ...process, children } : process,
+  );
+}
+
+function runtimeChildren(runtime: RuntimeStatusSnapshot): RuntimeChildInfo[] {
+  const processChildren = runtime.lifecycle.processes.map((process) => ({
+    name: process.name,
+    label: displayRuntimeName(process.name),
+    kind: "process" as const,
+    status: processStatusFromRuntimeProcess(process.status),
+    pid: process.pid,
+    exitCode: process.exit_code,
+  }));
+
+  const connectionChildren = runtime.lifecycle.connections.map((connection) => ({
+    name: connection.name,
+    label: displayRuntimeName(connection.name),
+    kind: "connection" as const,
+    status: processStatusFromReadiness(connection.readiness),
+    readiness: connection.readiness,
+  }));
+
+  return [...processChildren, ...connectionChildren];
+}
+
+function processStatusFromRuntimeProcess(
+  status: "running" | "exited" | "failed",
+): ProcessStatus {
+  if (status === "running") return "running";
+  if (status === "failed") return "errored";
+  return "stopped";
+}
+
+function processStatusFromReadiness(readiness: RuntimeReadiness): ProcessStatus {
+  if (readiness === "ready") return "running";
+  if (readiness === "starting") return "starting";
+  if (readiness === "degraded") return "errored";
+  return "stopped";
+}
+
+function displayRuntimeName(name: string): string {
+  return name
+    .split("_")
+    .filter(Boolean)
+    .map((part) => part[0]?.toUpperCase() + part.slice(1))
+    .join(" ");
 }
 
 function processAction(name: string, action: "start" | "stop" | "restart"): void {
