+++ console/vite.config.ts
@@ -18,6 +18,9 @@ export default defineConfig({
       "/health": {
         target: "http://127.0.0.1:8080",
       },
+      "/runtime": {
+        target: "http://127.0.0.1:8080",
+      },
     },
   },
 });
